Cleared the duplicate files
First time when I uploaded the files, I did it inside another folder. And it didn't let us get a github link.

In order to get the gihub link we had to upload all the files again in github.

Now I just removed the folder that contained all the files as we uploaded them again.